// ui-packages/ui-components/lib/essential/app-lab-button/appLabButton.type.ts
export enum ButtonType {
  Primary = 'primary',
  Secondary = 'secondary',
  Tertiary = 'tertiary',
}

export enum ButtonVariant {
  Action = 'action',
  Destructive = 'destructive',
  LowContrast = 'low-contrast',
}

export enum ButtonSize {
  XXSmall = 'xx-small',
  XSmall = 'x-small',
  Small = 'small',
  Large = 'large',
}